refactor(prisma): remove `helper.ts` and create `CsvParser` helper
